add project/Dependencies.scala
